refactor: move consent configs to a separate namespace
