+++ VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/Security/Configurations/IConsentProviderConfig.cs
@@ -1,6 +1,6 @@
 using Microsoft.IdentityModel.Tokens;
 
-namespace VirtualFinland.UserAPI.Security.Features;
+namespace VirtualFinland.UserAPI.Security.Configurations;
 
 public interface IConsentProviderConfig
 {
